Updater: Change the update jitter to a double instead of int.
Bug: 1144151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2640732
Commit-Queue: Mila Green <milagreen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845437}

+++ chrome/updater/external_constants_builder.h
@@ -36,7 +36,7 @@ class ExternalConstantsBuilder {
   ExternalConstantsBuilder& SetUseCUP(bool use_cup);
   ExternalConstantsBuilder& ClearUseCUP();
 
-  ExternalConstantsBuilder& SetInitialDelay(int initial_delay);
+  ExternalConstantsBuilder& SetInitialDelay(double initial_delay);
   ExternalConstantsBuilder& ClearInitialDelay();
 
   // Write the external constants overrides file in the default location
